(refactor) Create collections and call db from server.js, awaiting testing
