+++ server/db/collections/arcs.js
@@ -1,0 +1,8 @@
+var db = require('../config.js');
+var Arc = require('../models/arc.js');
+
+var Arcs = new db.Collection();
+
+Arcs.model = Arc;
+
+module.exports = Arcs;
