+++ server/db/collections/users.js
@@ -1,0 +1,8 @@
+var db = require('../config.js');
+var User = require('../models/user.js');
+
+var Users = new db.Collection();
+
+Users.model = User;
+
+module.exports = Users;
