+++ server/db/models/image.js
@@ -1,0 +1,12 @@
+var db = require('../config.js');
+var Arc = require('arc.js');
+
+var Image = db.Model.extend({
+  tableName: 'images',
+
+  arc: function () {
+    return this.belongsTo(Arc);
+  }
+});
+
+module.exports = Image;
