--- server/db/models/images.js
@@ -1,11 +1,0 @@
-var bookshelf = require('../config.js');
-
-var Image = db.Model.extend({
-  tableName: 'images',
-
-  initialize: function () {
-    
-  }
-});
-
-module.exports = Image;
