+++ server/db/models/user.js
@@ -1,14 +1,11 @@
-var bookshelf = require('../config.js');
+var db = require('../config.js');
+var Arc = require('arc.js');
 
 var User = db.Model.extend({
   tableName: 'users',
 
   arcs: function () {
     return this.hasMany(Arc);
-  },
-
-  initialize: function () {
-    
   }
 });
